backend/pyproject.toml
[project]
name = "ai-workspace-backend"
version = "0.3.0"
requires-python = ">=3.11"

dependencies = [
    # Core frameworks
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",   # ChatLiteLLM for library mode

    # LangGraph persistence
    "langgraph-checkpoint-postgres>=2.0.0",

    # LiteLLM — library mode (no proxy container in production)
    "litellm>=1.40.0",

    # LlamaIndex RAG
    "llama-index>=0.11.0",
    "llama-index-vector-stores-postgres>=0.2.0",
    "llama-index-embeddings-openai>=0.2.0",
    "llama-index-llms-openai>=0.2.0",
    "llama-index-llms-litellm>=0.1.0",

    # Hybrid retrieval — BM25 re-ranking
    "rank-bm25>=0.2.2",

    # FastAPI server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",

    # Database
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",

    # Redis — async client for rate limiting + Celery broker
    "redis>=5.0.0",

    # Celery background worker
    "celery[redis]>=5.3.0",

    # Auth — password hashing + JWT
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",

    # Schema migrations
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",

    # Structured logging
    "structlog>=24.0.0",

    # Config & utilities
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "flower>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
